+++ src/alignment/types/std_traits.rs
@@ -1,6 +1,5 @@
-use super::AlignmentStates;
 use crate::{
-    alignment::{Alignment, MaybeAligned},
+    alignment::{Alignment, MaybeAligned, types::AlignmentStates},
     data::{
         cigar::{CigarView, is_valid_op},
         types::cigar::{Cigar, CigarError, Ciglet, CigletIteratorChecked},
@@ -77,9 +76,15 @@ impl FromIterator<Ciglet> for Result<AlignmentStates, CigarError> {
             // The current index being accumulated into
             let mut merged_index = 0;
             while let Some(ciglet) = out.get(original_index).copied() {
-                if ciglet.op == out[merged_index].op {
-                    out[merged_index].inc += ciglet.inc;
+                let merged_ciglet = &mut out[merged_index];
+
+                if ciglet.op == merged_ciglet.op {
+                    merged_ciglet.inc = merged_ciglet
+                        .inc
+                        .checked_add(ciglet.inc)
+                        .ok_or(CigarError::MergeIncOverflow)?;
                 } else {
+                    // Advance to the next available position to accumulate into
                     merged_index += 1;
                     out[merged_index] = out[original_index];
                 }
@@ -114,18 +119,6 @@ impl std::fmt::Display for AlignmentStates {
 impl TryFrom<&[u8]> for AlignmentStates {
     type Error = CigarError;
 
-    /// Tries to convert a byte slice into an [`AlignmentStates`].
-    ///
-    /// Note that this will allocate a new vector. `*` is converted to an empty
-    /// alignment states struct.
-    ///
-    /// ## Errors
-    ///
-    /// - The CIGAR operations must be among `M, I, D, N, S, H, P, X, =`
-    /// - Every operation in the CIGAR string must have a preceding increment
-    /// - Every increment must be followed by an operation
-    /// - The increment for each operation must be non-zero and less than or
-    ///   equal to [`usize::MAX`]
     fn try_from(v: &[u8]) -> Result<Self, CigarError> {
         let mut states = AlignmentStates::new();
 
@@ -144,7 +137,7 @@ impl TryFrom<&[u8]> for AlignmentStates {
             for ciglet in iter {
                 let ciglet = ciglet?;
                 if prev_ciglet.op == ciglet.op {
-                    prev_ciglet.inc += ciglet.inc;
+                    prev_ciglet.inc = prev_ciglet.inc.checked_add(ciglet.inc).ok_or(CigarError::MergeIncOverflow)?;
                 } else {
                     // Validity: CigletIteratorChecked errors on zero
                     // increments, and we have ensured the operation is
@@ -164,18 +157,6 @@ impl TryFrom<&[u8]> for AlignmentStates {
 impl TryFrom<String> for AlignmentStates {
     type Error = CigarError;
 
-    /// Tries to convert a string into [`AlignmentStates`].
-    ///
-    /// Note that this will allocate a new vector. `*` is converted to an empty
-    /// alignment states struct.
-    ///
-    /// ## Errors
-    ///
-    /// - The CIGAR operations must be among `M, I, D, N, S, H, P, X, =`
-    /// - Every operation in the CIGAR string must have a preceding increment
-    /// - Every increment must be followed by an operation
-    /// - The increment for each operation must be non-zero and less than or
-    ///   equal to [`usize::MAX`]
     #[inline]
     fn try_from(s: String) -> Result<Self, Self::Error> {
         AlignmentStates::try_from(s.as_bytes())
@@ -185,18 +166,6 @@ impl TryFrom<String> for AlignmentStates {
 impl TryFrom<Vec<u8>> for AlignmentStates {
     type Error = CigarError;
 
-    /// Tries to convert a vector of bytes into [`AlignmentStates`].
-    ///
-    /// Note that this will allocate a new vector. `*` is converted to an empty
-    /// alignment states struct.
-    ///
-    /// ## Errors
-    ///
-    /// - The CIGAR operations must be among `M, I, D, N, S, H, P, X, =`
-    /// - Every operation in the CIGAR string must have a preceding increment
-    /// - Every increment must be followed by an operation
-    /// - The increment for each operation must be non-zero and less than or
-    ///   equal to [`usize::MAX`]
     #[inline]
     fn try_from(bytes: Vec<u8>) -> Result<Self, Self::Error> {
         AlignmentStates::try_from(bytes.as_slice())
@@ -206,18 +175,6 @@ impl TryFrom<Vec<u8>> for AlignmentStates {
 impl<const N: usize> TryFrom<[u8; N]> for AlignmentStates {
     type Error = CigarError;
 
-    /// Tries to convert an array of bytes into [`AlignmentStates`].
-    ///
-    /// Note that this will allocate a new vector. `*` is converted to an empty
-    /// alignment states struct.
-    ///
-    /// ## Errors
-    ///
-    /// - The CIGAR operations must be among `M, I, D, N, S, H, P, X, =`
-    /// - Every operation in the CIGAR string must have a preceding increment
-    /// - Every increment must be followed by an operation
-    /// - The increment for each operation must be non-zero and less than or
-    ///   equal to [`usize::MAX`]
     #[inline]
     fn try_from(v: [u8; N]) -> Result<Self, Self::Error> {
         AlignmentStates::try_from(v.as_slice())
@@ -227,18 +184,6 @@ impl<const N: usize> TryFrom<[u8; N]> for AlignmentStates {
 impl TryFrom<&mut [u8]> for AlignmentStates {
     type Error = CigarError;
 
-    /// Tries to convert a mutable slice of bytes into [`AlignmentStates`].
-    ///
-    /// Note that this will allocate a new vector. `*` is converted to an empty
-    /// alignment states struct.
-    ///
-    /// ## Errors
-    ///
-    /// - The CIGAR operations must be among `M, I, D, N, S, H, P, X, =`
-    /// - Every operation in the CIGAR string must have a preceding increment
-    /// - Every increment must be followed by an operation
-    /// - The increment for each operation must be non-zero and less than or
-    ///   equal to [`usize::MAX`]
     #[inline]
     fn try_from(v: &mut [u8]) -> Result<Self, Self::Error> {
         AlignmentStates::try_from(&*v)
@@ -248,18 +193,6 @@ impl TryFrom<&mut [u8]> for AlignmentStates {
 impl<const N: usize> TryFrom<&[u8; N]> for AlignmentStates {
     type Error = CigarError;
 
-    /// Tries to convert an array of bytes into [`AlignmentStates`].
-    ///
-    /// Note that this will allocate a new vector. `*` is converted to an empty
-    /// alignment states struct.
-    ///
-    /// ## Errors
-    ///
-    /// - The CIGAR operations must be among `M, I, D, N, S, H, P, X, =`
-    /// - Every operation in the CIGAR string must have a preceding increment
-    /// - Every increment must be followed by an operation
-    /// - The increment for each operation must be non-zero and less than or
-    ///   equal to [`usize::MAX`]
     #[inline]
     fn try_from(v: &[u8; N]) -> Result<Self, Self::Error> {
         AlignmentStates::try_from(v.as_slice())
@@ -269,18 +202,6 @@ impl<const N: usize> TryFrom<&[u8; N]> for AlignmentStates {
 impl<const N: usize> TryFrom<&mut [u8; N]> for AlignmentStates {
     type Error = CigarError;
 
-    /// Tries to convert an array of bytes into [`AlignmentStates`].
-    ///
-    /// Note that this will allocate a new vector. `*` is converted to an empty
-    /// alignment states struct.
-    ///
-    /// ## Errors
-    ///
-    /// - The CIGAR operations must be among `M, I, D, N, S, H, P, X, =`
-    /// - Every operation in the CIGAR string must have a preceding increment
-    /// - Every increment must be followed by an operation
-    /// - The increment for each operation must be non-zero and less than or
-    ///   equal to [`usize::MAX`]
     #[inline]
     fn try_from(v: &mut [u8; N]) -> Result<Self, Self::Error> {
         AlignmentStates::try_from(v.as_slice())
@@ -290,18 +211,6 @@ impl<const N: usize> TryFrom<&mut [u8; N]> for AlignmentStates {
 impl TryFrom<&str> for AlignmentStates {
     type Error = CigarError;
 
-    /// Tries to convert a string slice into [`AlignmentStates`].
-    ///
-    /// Note that this will allocate a new vector. `*` is converted to an empty
-    /// alignment states struct.
-    ///
-    /// ## Errors
-    ///
-    /// - The CIGAR operations must be among `M, I, D, N, S, H, P, X, =`
-    /// - Every operation in the CIGAR string must have a preceding increment
-    /// - Every increment must be followed by an operation
-    /// - The increment for each operation must be non-zero and less than or
-    ///   equal to [`usize::MAX`]
     #[inline]
     fn try_from(s: &str) -> Result<Self, Self::Error> {
         AlignmentStates::try_from(s.as_bytes())
@@ -311,18 +220,6 @@ impl TryFrom<&str> for AlignmentStates {
 impl TryFrom<Cigar> for AlignmentStates {
     type Error = CigarError;
 
-    /// Tries to convert a CIGAR string into [`AlignmentStates`].
-    ///
-    /// Note that this will allocate a new vector. `*` is converted to an empty
-    /// alignment states struct.
-    ///
-    /// ## Errors
-    ///
-    /// - The CIGAR operations must be among `M, I, D, N, S, H, P, X, =`
-    /// - Every operation in the CIGAR string must have a preceding increment
-    /// - Every increment must be followed by an operation
-    /// - The increment for each operation must be non-zero and less than or
-    ///   equal to [`usize::MAX`]
     #[inline]
     fn try_from(value: Cigar) -> Result<Self, Self::Error> {
         AlignmentStates::try_from(value.as_bytes())
@@ -332,18 +229,6 @@ impl TryFrom<Cigar> for AlignmentStates {
 impl TryFrom<&Cigar> for AlignmentStates {
     type Error = CigarError;
 
-    /// Tries to convert a CIGAR string into [`AlignmentStates`].
-    ///
-    /// Note that this will allocate a new vector. `*` is converted to an empty
-    /// alignment states struct.
-    ///
-    /// ## Errors
-    ///
-    /// - The CIGAR operations must be among `M, I, D, N, S, H, P, X, =`
-    /// - Every operation in the CIGAR string must have a preceding increment
-    /// - Every increment must be followed by an operation
-    /// - The increment for each operation must be non-zero and less than or
-    ///   equal to [`usize::MAX`]
     #[inline]
     fn try_from(value: &Cigar) -> Result<Self, Self::Error> {
         AlignmentStates::try_from(value.as_bytes())
@@ -353,18 +238,6 @@ impl TryFrom<&Cigar> for AlignmentStates {
 impl<'a> TryFrom<CigarView<'a>> for AlignmentStates {
     type Error = CigarError;
 
-    /// Tries to convert a CIGAR string into [`AlignmentStates`].
-    ///
-    /// Note that this will allocate a new vector. `*` is converted to an empty
-    /// alignment states struct.
-    ///
-    /// ## Errors
-    ///
-    /// - The CIGAR operations must be among `M, I, D, N, S, H, P, X, =`
-    /// - Every operation in the CIGAR string must have a preceding increment
-    /// - Every increment must be followed by an operation
-    /// - The increment for each operation must be non-zero and less than or
-    ///   equal to [`usize::MAX`]
     #[inline]
     fn try_from(value: CigarView<'a>) -> Result<Self, Self::Error> {
         AlignmentStates::try_from(value.as_bytes())