+++ src/data/types/cigar/error.rs
@@ -7,15 +7,18 @@ use std::{error::Error, fmt};
 #[derive(Clone, Eq, PartialEq, Copy, Hash)]
 /// Custom error type for constructing CIGAR strings.
 pub enum CigarError {
-    /// The CIGAR operation must be one of: `M, I, D, N, S, H, P, X, =`
+    /// The CIGAR operation must be one of: `M, I, D, N, S, H, P, X, =`.
     InvalidOperation,
-    /// The CIGAR increment must be a non-zero positive integer
+    /// The CIGAR increment must be a non-zero positive integer.
     IncZero,
-    /// The CIGAR increment must be less than or equal to [`usize::MAX`]
+    /// The CIGAR increment must be less than or equal to [`usize::MAX`].
     IncOverflow,
-    /// The CIGAR operation must have preceding increment
+    /// The CIGAR increment after merging consecutive ciglets with the same
+    /// operation must be less than or equal to [`usize::MAX`].
+    MergeIncOverflow,
+    /// The CIGAR operation must have preceding increment.
     MissingInc,
-    /// The CIGAR increment must be followed by operation
+    /// The CIGAR increment must be followed by operation.
     MissingOp,
 }
 
@@ -26,6 +29,11 @@ impl fmt::Display for CigarError {
             CigarError::InvalidOperation => f.write_str("CIGAR operation must be one of: M, I, D, N, S, H, P, X, ="),
             CigarError::IncZero => f.write_str("CIGAR increment must be a non-zero positive integer"),
             CigarError::IncOverflow => write!(f, "CIGAR increment must be less than or equal to {}", usize::MAX),
+            CigarError::MergeIncOverflow => write!(
+                f,
+                "After merging consecutive ciglets with the same operation, the total increment must be less than or equal to {}",
+                usize::MAX
+            ),
             CigarError::MissingInc => f.write_str("CIGAR operation must have preceding increment"),
             CigarError::MissingOp => f.write_str("CIGAR increment must be followed by operation"),
         }